refactor: combat.py에서 CombatManager를 combat_manager.py로 분리 (태스크 29.5)
- combat.py: enum/dataclass만 유지 (CombatAction, CombatantType, Combatant, CombatTurn, CombatInstance)
- combat_manager.py: CombatManager 클래스 분리
- game_engine.py, combat_handler.py import 경로 수정
