+++ src/mud_engine/core/game_engine.py
@@ -59,7 +59,7 @@ class GameEngine:
         self.world_manager.set_game_engine(self)
 
         # CombatManager 및 CombatHandler 초기화
-        from ..game.combat import CombatManager
+        from ..game.combat_manager import CombatManager
         from ..game.combat_handler import CombatHandler
         self.combat_manager = CombatManager(self.session_manager)
         self.combat_handler = CombatHandler(self.combat_manager, self.world_manager, self, self.session_manager)
